// src/api/endpoints/model/sysRole.ts
/**
 * Generated by orval v7.21.0 🍺
 * Do not edit manually.
 * (title)
 * OpenAPI spec version: 0.0.0
 */

export interface SysRole {
  id: string;
  tenantId: string;
  clientId: string;
  /**
   * @minLength 1
   * @maxLength 64
   */
  roleCode: string;
  /**
   * @minLength 1
   * @maxLength 64
   */
  roleName: string;
  description?: string;
  isPreset: boolean;
  status: number;
  createdAt: string;
  updatedAt: string;
}